+++ algorithms/linfa-pls/src/lib.rs
@@ -1,4 +1,5 @@
 mod algorithm;
 pub mod errors;
+mod utils;
 
 pub use algorithm::*;
